separate servers for llamacpp and hf
